whitelist script tag inserted for test file loaded assertion using nonce
This adds a nonce to <script> tag inserted by ember-cli for assertion
that test file was loaded. Otherwise this violates default CSP in tests.

Nonce is always present in CSP header but only part of CSP meta tag
injected into `tests/index.html`.

A nonce used in productive applications must be cryptographically random.
But since this one is only used in local development and if executing
tests, using a static string is fine.