Set status-code to 204 (no content)
With Cross-Origin Read Blocking (CORB), the JSON response
we've used to send as a reply to the CSP-report, is hit with
CORB blocking.

While this is harmless (because the client doesn't
care about the response to the CSP report), we want to avoid the
warning that is shown in Chrome, and possibly other browsers.

By not sending any response body, we avoid triggering CORB-blocking.