do not set X-Content-Security-Policy header
All major browsers support Content-Security-Policy header since a long time.
Only Internet Explorer 11 does not support it. But it also only supports
`sandbox` directive by X-Content-Security-Policy header. The sandbox is not
useful at all for Ember applications cause it could only be used to disable
all scripts. As an Ember applicaiton is not usable at all without JavaScript
that is not a reasonable configuration.